Remove old cronjob controller
